requirements.txt
anyio==3.6.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
CoolProp==6.4.1
decorator==5.1.1
Deprecated==1.2.13
docker==6.0.0
executing==1.0.0
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==1.10.1
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
python-decouple==3.6
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
stack-data==0.5.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
websocket-client==1.4.1
wrapt==1.14.1
